+ Extract accessors to concern
